image-builder and image-xlnx-xen: Fix linux-xlnx do_deploy errors
When the system is configured with:

  INHERIT += "plnx-deploy"
  IMAGE_PLNX_XEN_DEPLOY = "1"
  INITRAMFS_IMAGE ?= "petalinux-initramfs-image"
  INITRAMFS_FSTYPES ?= "cpio.gz cpio.gz.u-boot tar.gz"

and the user builds the linux-xlnx recipe, errors can occur:

  DEBUG: Executing shell function plnx_compile_image_builder
  .../deploy/images/xilinx-vck190/uboot-script-gen: line 1100: .../deploy/images/xilinx-vck190/common: No such file or directory
  File xen doesn't exist, exiting
  .../deploy/images/xilinx-vck190/uboot-script-gen: line 10: remove_tmp_files: command not found
  WARNING: .../work/xilinx_vck190-xilinx-linux/linux-xlnx/6.12.0+git/temp/run.plnx_compile_image_builder.2756700:153 exit 1 from '.../deploy/images/xilinx-vck190/uboot-script-gen -c /proj/petalinux/2025.1_next/petalinux-v2025.1_03052037_dev/logs/bsps/xilinx-vck190-xsct-2025.1/pre-built/linux/xen/xen.cfg -t tftp -d $(dirname "/proj/petalinux/2025.1_next/petalinux-v2025.1_03052037_dev/logs/bsps/xilinx-vck190-xsct-2025.1/pre-built/linux/xen/xen.cfg") -o /proj/petalinux/2025.1_next/petalinux-v2025.1_03052037_dev/logs/bsps/xilinx-vck190-xsct-2025.1/images/linux/xen_boot_tftp'

The fix for the above is in two (related) parts:

image-builder do_deploy:class-native was only installing the script, but not
the ancillary files, such as 'common' that it was looking for

The image-xlnx-xen.bbclass needs to be able to call the script from within
the directory where all of the components exist, not just the main one.

Additionally the 'rootfs.cpio.gz' file was being copied into the WORKDIR, but
was being expected to be in the directory holding 'XEN_CONFIG' according to
the call.  Note, this is not really a best practice as we're now building
outside of the WORKDIR, but the Xen implementation will be changing in the
future so this behavior is really PetaLinux specific.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>